Fix: beberapa major bug pembelian, perbesar range grandtotal, dan minor bug/error

+++ database/migrations/2021_10_15_164649_create_adjustment_plus_table.php
@@ -19,8 +19,8 @@ class CreateAdjustmentPlusTable extends Migration
             $table->date('tanggal');
             $table->foreignId('gudang_id')->constrained('gudang');
             $table->foreignId('matauang_id')->constrained('matauang');
-            $table->float('rate');
-            $table->float('grand_total');
+            $table->double('rate');
+            $table->double('grand_total', 20, 2);
             $table->timestamps();
         });
     }
